refactor: rename quote files to be correct with abstractions
