+++ router/usecase/quote_out_given_in.go
@@ -2,9 +2,11 @@ package usecase
 
 import (
 	"context"
+	"fmt"
+	"strings"
 
 	"github.com/osmosis-labs/sqs/domain"
-	"github.com/osmosis-labs/sqs/router/types"
+	"github.com/osmosis-labs/sqs/router/usecase/route"
 
 	"github.com/osmosis-labs/osmosis/osmomath"
 
@@ -12,15 +14,29 @@ import (
 )
 
 var (
-	_ domain.Quote = &quoteExactAmountOut{}
+	one = osmomath.OneDec()
 )
 
-// quoteExactAmountOut is a quote wrapper for exact out quotes.
-// Note that only the PrepareResult method is different from the quoteExactAmountIn.
-type quoteExactAmountOut struct {
-	*quoteExactAmountIn     "json:\"-\""
-	AmountIn                osmomath.Int        "json:\"amount_in\""
-	AmountOut               sdk.Coin            "json:\"amount_out\""
+var (
+	_ domain.Quote = &quoteExactAmountIn{}
+)
+
+// NOTE: This is because structs in alias declaration are not exported
+type (
+	QuoteExactAmountOut = quoteExactAmountOut
+	QuoteExactAmountIn  = quoteExactAmountIn
+)
+
+func NewQuoteExactAmountOut(q *QuoteExactAmountIn) *quoteExactAmountOut {
+	return &quoteExactAmountOut{
+		quoteExactAmountIn: q,
+	}
+}
+
+// quoteExactAmountIn is a quote implementation for token swap method exact in.
+type quoteExactAmountIn struct {
+	AmountIn                sdk.Coin            "json:\"amount_in\""
+	AmountOut               osmomath.Int        "json:\"amount_out\""
 	Route                   []domain.SplitRoute "json:\"route\""
 	EffectiveFee            osmomath.Dec        "json:\"effective_fee\""
 	PriceImpact             osmomath.Dec        "json:\"price_impact\""
@@ -35,37 +51,102 @@ type quoteExactAmountOut struct {
 // Computes an effective spread factor from all routes.
 //
 // Returns the updated route and the effective spread factor.
-func (q *quoteExactAmountOut) PrepareResult(ctx context.Context, scalingFactor osmomath.Dec) ([]domain.SplitRoute, osmomath.Dec, error) {
-	// Prepare exact out in the quote for inputs inversion
-	if _, _, err := q.quoteExactAmountIn.PrepareResult(ctx, scalingFactor); err != nil {
-		return nil, osmomath.Dec{}, err
-	}
+func (q *quoteExactAmountIn) PrepareResult(ctx context.Context, scalingFactor osmomath.Dec) ([]domain.SplitRoute, osmomath.Dec, error) {
+	totalAmountIn := q.AmountIn.Amount.ToLegacyDec()
+	totalFeeAcrossRoutes := osmomath.ZeroDec()
 
-	// Assign the inverted values to the quote
-	q.AmountOut = q.quoteExactAmountIn.AmountIn
-	q.AmountIn = q.quoteExactAmountIn.AmountOut
-	q.Route = q.quoteExactAmountIn.Route
-	q.EffectiveFee = q.quoteExactAmountIn.EffectiveFee
-	q.PriceImpact = q.quoteExactAmountIn.PriceImpact
-	q.InBaseOutQuoteSpotPrice = q.quoteExactAmountIn.InBaseOutQuoteSpotPrice
+	totalSpotPriceInBaseOutQuote := osmomath.ZeroDec()
+	totalEffectiveSpotPriceInBaseOutQuote := osmomath.ZeroDec()
 
-	for i, route := range q.Route {
-		route, ok := route.(*RouteWithOutAmount)
-		if !ok {
-			return nil, osmomath.Dec{}, types.ErrInvalidRouteType
+	resultRoutes := make([]domain.SplitRoute, 0, len(q.Route))
+
+	for _, curRoute := range q.Route {
+		routeTotalFee := osmomath.ZeroDec()
+		routeAmountInFraction := curRoute.GetAmountIn().ToLegacyDec().Quo(totalAmountIn)
+
+		// Calculate the spread factor across pools in the route
+		for _, pool := range curRoute.GetPools() {
+			poolTakerFee := pool.GetTakerFee()
+
+			routeTotalFee.AddMut(
+				//  (1 - routeTotalFee) * poolTakerFee
+				osmomath.OneDec().SubMut(routeTotalFee).MulTruncateMut(poolTakerFee),
+			)
 		}
 
-		// invert the in and out amounts
-		route.InAmount, route.OutAmount = route.OutAmount, route.InAmount
+		// Update the spread factor pro-rated by the amount in
+		totalFeeAcrossRoutes.AddMut(routeTotalFee.MulMut(routeAmountInFraction))
 
-		q.Route[i] = route
-
-		// invert the in and out amounts for each pool
-		for _, p := range route.GetPools() {
-			p.SetTokenInDenom(p.GetTokenOutDenom())
-			p.SetTokenOutDenom("")
+		amountInFraction := q.AmountIn.Amount.ToLegacyDec().MulMut(routeAmountInFraction).TruncateInt()
+		newPools, routeSpotPriceInBaseOutQuote, effectiveSpotPriceInBaseOutQuote, err := curRoute.PrepareResultPools(ctx, sdk.NewCoin(q.AmountIn.Denom, amountInFraction))
+		if err != nil {
+			return nil, osmomath.Dec{}, err
 		}
+
+		totalSpotPriceInBaseOutQuote = totalSpotPriceInBaseOutQuote.AddMut(routeSpotPriceInBaseOutQuote.MulMut(routeAmountInFraction))
+		totalEffectiveSpotPriceInBaseOutQuote = totalEffectiveSpotPriceInBaseOutQuote.AddMut(effectiveSpotPriceInBaseOutQuote.MulMut(routeAmountInFraction))
+
+		resultRoutes = append(resultRoutes, &RouteWithOutAmount{
+			RouteImpl: route.RouteImpl{
+				Pools:                      newPools,
+				HasGeneralizedCosmWasmPool: curRoute.ContainsGeneralizedCosmWasmPool(),
+			},
+			InAmount:  curRoute.GetAmountIn(),
+			OutAmount: curRoute.GetAmountOut(),
+		})
 	}
+
+	// Calculate price impact
+	if !totalSpotPriceInBaseOutQuote.IsZero() {
+		q.PriceImpact = totalEffectiveSpotPriceInBaseOutQuote.Quo(totalSpotPriceInBaseOutQuote).SubMut(one)
+	}
+
+	q.EffectiveFee = totalFeeAcrossRoutes
+	q.Route = resultRoutes
+	q.InBaseOutQuoteSpotPrice = totalSpotPriceInBaseOutQuote
 
 	return q.Route, q.EffectiveFee, nil
+}
+
+// GetAmountIn implements Quote.
+func (q *quoteExactAmountIn) GetAmountIn() sdk.Coin {
+	return q.AmountIn
+}
+
+// GetAmountOut implements Quote.
+func (q *quoteExactAmountIn) GetAmountOut() osmomath.Int {
+	return q.AmountOut
+}
+
+// GetRoute implements Quote.
+func (q *quoteExactAmountIn) GetRoute() []domain.SplitRoute {
+	return q.Route
+}
+
+// GetEffectiveFee implements Quote.
+func (q *quoteExactAmountIn) GetEffectiveFee() osmomath.Dec {
+	return q.EffectiveFee
+}
+
+// String implements domain.Quote.
+func (q *quoteExactAmountIn) String() string {
+	var builder strings.Builder
+
+	builder.WriteString(fmt.Sprintf("Quote: %s in for %s out \n", q.AmountIn, q.AmountOut))
+
+	for _, route := range q.Route {
+		builder.WriteString(route.String())
+	}
+
+	return builder.String()
+}
+
+// GetPriceImpact implements domain.Quote.
+func (q *quoteExactAmountIn) GetPriceImpact() osmomath.Dec {
+	return q.PriceImpact
+}
+
+// GetInBaseOutQuoteSpotPrice implements domain.Quote.
+func (q *quoteExactAmountIn) GetInBaseOutQuoteSpotPrice() osmomath.Dec {
+	return q.InBaseOutQuoteSpotPrice
 }
